Apply theme colors to TopAppBar and remove LazyListState
This commit updates the `DiffCheckerScreen.kt` to apply theme colors to the `TopAppBar`. Specifically, the `containerColor` is set to `Color.White` and `titleContentColor` to `primary`.

Additionally, the `rememberLazyListState` instances (`scrollStateOld` and `scrollStateNew`) have been removed from `CharDiffText.kt` as they were not being used to synchronize scrolling between the two `LazyColumn`s.